Fixes #7: allow for mandatory authentication for Music Assistant 2.7+
